code for opening browser the browser has been added
