+++ src/main/java/utils/commonMethods.java
@@ -5,30 +5,8 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 
 public class commonMethods {
-    @FindBy(id="txtUsername")
-    public WebElement usernameBox;
-
-    @FindBy (name="txtPassword")
-    public WebElement passwordBox;
-
-    @FindBy (id="btnLogin")
-    public WebElement loginBtn;
-
-    @FindBy (xpath="//*[@id=\"divLogo\"]/img")
-    public WebElement logo;
-
-    @FindBy ( id="spanMessage")
-    public WebElement errorMssage;
-
-    public LoginPage(){
-        PageFactory.initElements(driver, this);
-    }
-    public  void LoginMethod(String username , String password){
-        sendText(usernameBox,username);
-        sendText(passwordBox,password);
-        loginBtn.click();
-
+   //code for opening the browser
     }
 
 
-}
+
